+++ stroom-app/src/test/java/stroom/test/ContentImportService.java
@@ -1,5 +1,7 @@
 package stroom.test;
 
+import stroom.content.ContentPack;
+import stroom.content.ContentPacks;
 import stroom.importexport.impl.ImportExportService;
 import stroom.test.common.StroomCoreServerTestFileUtil;
 import stroom.util.io.FileUtil;
@@ -24,39 +26,6 @@ public class ContentImportService {
 
     public static final String CONTENT_PACK_IMPORT_DIR = "transientContentPacks";
 
-    public enum ContentPackName {
-        CORE_XML_SCHEMAS("core-xml-schemas"),
-        EVENT_LOGGING_XML_SCHEMA("event-logging-xml-schema"),
-        TEMPLATE_PIPELINES("template-pipelines"),
-        STANDARD_PIPELINES("standard-pipelines");
-
-        private final String packName;
-
-        ContentPackName(final String packName) {
-            this.packName = packName;
-        }
-
-        String getPackName() {
-            return packName;
-        }
-    }
-
-    public static final ContentPack CORE_XML_SCHEMAS_PACK = ContentPack.of(
-            ContentPackName.CORE_XML_SCHEMAS,
-            Version.of(1, 1));
-
-    public static final ContentPack EVENT_LOGGING_XML_SCHEMA_PACK = ContentPack.of(
-            ContentPackName.EVENT_LOGGING_XML_SCHEMA,
-            Version.of(3, 1, 1));
-
-    public static final ContentPack TEMPLATE_PIPELINES_PACK = ContentPack.of(
-            ContentPackName.TEMPLATE_PIPELINES,
-            Version.of(0, 2));
-
-    public static final ContentPack STANDARD_PIPELINES_PACK = ContentPack.of(
-            ContentPackName.STANDARD_PIPELINES,
-            Version.of(0, 1));
-
     private static final Version VISUALISATIONS_VERSION = Version.of(3, 0, 4);
 
     private ImportExportService importExportService;
@@ -71,10 +40,10 @@ public class ContentImportService {
      */
     public void importStandardPacks() {
         importContentPacks(Arrays.asList(
-                CORE_XML_SCHEMAS_PACK,
-                EVENT_LOGGING_XML_SCHEMA_PACK,
-                TEMPLATE_PIPELINES_PACK,
-                STANDARD_PIPELINES_PACK
+                ContentPacks.CORE_XML_SCHEMAS_PACK,
+                ContentPacks.EVENT_LOGGING_XML_SCHEMA_PACK,
+                ContentPacks.TEMPLATE_PIPELINES_PACK,
+                ContentPacks.STANDARD_PIPELINES_PACK
         ));
     }
 
@@ -107,29 +76,4 @@ public class ContentImportService {
         return contentPackDir;
     }
 
-    public static class ContentPack {
-        private final ContentPackName name;
-        private final Version version;
-
-        public ContentPack(final ContentPackName name, final Version version) {
-            this.name = name;
-            this.version = version;
-        }
-
-        public static ContentPack of(final ContentPackName name, final Version version) {
-            return new ContentPack(name, version);
-        }
-
-        public String getNameAsStr() {
-            return name.getPackName();
-        }
-
-        ContentPackName getName() {
-            return name;
-        }
-
-        public Version getVersion() {
-            return version;
-        }
-    }
 }