+++ stroom-pipeline/src/test/java/stroom/pipeline/refdata/store/offheapstore/serdes/TestUnsignedLongSerde.java
@@ -1,0 +1,73 @@
+/*
+ * Copyright 2018 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package stroom.pipeline.refdata.store.offheapstore.serdes;
+
+
+import stroom.pipeline.refdata.store.offheapstore.UnsignedBytesInstances;
+import stroom.pipeline.refdata.store.offheapstore.UnsignedLong;
+
+import org.junit.jupiter.api.DynamicTest;
+import org.junit.jupiter.api.TestFactory;
+
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Supplier;
+import java.util.stream.IntStream;
+import java.util.stream.Stream;
+
+class TestUnsignedLongSerde extends AbstractSerdeTest<UnsignedLong, UnsignedLongSerde> {
+
+    private AtomicInteger length = new AtomicInteger(-1);
+
+//    @Test
+//    void testSerialisationDeserialisation() {
+//        final UnsignedLong unsignedLong = UnsignedLong.of(1L, length.get());
+//
+//        doSerialisationDeserialisationTest(unsignedLong);
+//    }
+
+    @TestFactory
+    Stream<DynamicTest> testSerialisationDeserialisation() {
+        return IntStream.rangeClosed(1, 8)
+                .boxed()
+                .map(len ->
+                        DynamicTest.dynamicTest("length " + len, () -> {
+                    length.set(len);
+                    long val = UnsignedBytesInstances.of(len).getMaxVal();
+                    final UnsignedLong unsignedLong = UnsignedLong.of(val, len);
+
+                    doSerialisationDeserialisationTest(unsignedLong);
+                }));
+    }
+
+    @Override
+    Class<UnsignedLongSerde> getSerdeType() {
+        return UnsignedLongSerde.class;
+    }
+
+    @Override
+    Supplier<UnsignedLongSerde> getSerdeSupplier() {
+        return () ->
+                new UnsignedLongSerde(length.get());
+    }
+
+    @Override
+    UnsignedLongSerde getSerde() {
+        // As we are using a TestFactory we need a fresh serde on each call to doSerialisationDeserialisationTest
+        return getSerdeSupplier().get();
+    }
+}